refactor: 이미 예약중인 테마 아이디를 삭제할 경우 예외가 발생하도록 수정

+++ src/main/java/roomescape/repository/impl/ThemeRepositoryImpl.java
@@ -3,17 +3,21 @@ package roomescape.repository.impl;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Component;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
+import roomescape.exception.InvalidReservationException;
 import roomescape.repository.ThemeRepository;
 
 @Component
 public class ThemeRepositoryImpl implements ThemeRepository {
 
     private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
 
-    public ThemeRepositoryImpl(ThemeDao themeDao) {
+    public ThemeRepositoryImpl(ThemeDao themeDao, ReservationDao reservationDao) {
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
 
     @Override
@@ -30,6 +34,9 @@ public class ThemeRepositoryImpl implements ThemeRepository {
     @Override
     public void delete(Long id) {
         findById(id);
+        if (reservationDao.countExistReservationByTheme(id) != 0) {
+            throw new InvalidReservationException("이미 예약된 테마를 삭제할 수 없습니다.");
+        }
         themeDao.delete(id);
     }
